+++ src/utils/common.py
@@ -168,6 +168,11 @@ def unhandled_exception():
     print(settings.print_critical_msg(err_msg))
     raise SystemExit()
 
+  elif "Invalid IPv6 URL" in exc_msg:
+    err_msg = "invalid URL ('" + exc_msg.strip().split('\n')[-1] + "')"
+    print(settings.print_critical_msg(err_msg))
+    raise SystemExit()
+
   elif any(_ in exc_msg for _ in ("The paging file is too small",)):
     err_msg = "No space left for paging file."
     print(settings.print_critical_msg(err_msg))