+++ src/thirdparty/flatten_json/flatten_json.py
@@ -94,7 +94,8 @@ def unflatten(flat_dict, separator='_'):
 
     def _unflatten(dic, keys, value):
         for key in keys[:-1]:
-            dic = dic.setdefault(key, {})
+            # dic = dic.setdefault(key, {})
+            dic = dic.setdefault(key, OrderedDict())
 
         dic[keys[-1]] = value